Improve accessibility menu component Uses teleport to move the dialog and accessibility button to the same component
