+++ src/components/icons/IconClose.vue
@@ -1,0 +1,8 @@
+<template>
+  <svg class="icon" viewBox="0 -960 960 960">
+    <path
+      fill="currentColor"
+      d="m256-200-56-56 224-224-224-224 56-56 224 224 224-224 56 56-224 224 224 224-56 56-224-224-224 224Z"
+    />
+  </svg>
+</template>
